Add (and allow reloading of) config template
Adds a basic collector.yaml template along with a +TARGETS file that
tells OPNsense where the file should live when it's installed/rebuilt,
something we tell it to do any time the form has been saved.

Also removes the requirement on the API Endpoint, adding instead a mask
to validate against, along with a vague message about what it expects.