Utilise template's apply button so it shows right
It was bugging me that the Save button wasn't part of the actual form
and wasn't spaced/aligned very well as a result, and nor could you
submit the form by pressing enter with an input focused.

By setting an ID we get a pre-built Apply button. It's still just a
basic button by default, but setting its type to submit can actually
solve that now since it's finally part of the form its dealing with.