Split/reorganise settings
Splits the version info "fields" out into their own "form" so they can
be separated in the view with a bit of space between the two sections,
and the added benefit of the 'full help' toggle being with the form.

Keeps the heading field for the Status form which makes it collapsible,
but removes it from the main settings form since it's a bit redundant,
and ultimately separates the form itself from the full help toggle,
which ends up looking rather odd. But that's probably just my OCD.

Also switches the debug/syslog fields and adds some missing help text.
Oh yeah, and strips the `settings.` key from the model so values can be
accessed with simply `OPNsense.Scrutiny.host_id` for example. That was
kinda the point of this commit initially. Distractions, distractions.